requirements.txt
# Core
python-dotenv>=1.0.0
pydantic>=2.0.0
pillow>=10.0.0

# File handling
python-magic>=0.4.27
pandas>=2.0.0
openpyxl>=3.1.0
pypdf>=4.0.0
pdf2image>=1.16.0

# OCR
pytesseract>=0.3.10
easyocr>=1.7.0

# Layout analysis
torch>=2.0.0
transformers>=4.30.0
ultralytics>=8.0.0

# Table extraction (pdfplumber remplace camelot)
pdfplumber>=0.10.0

# NLP
spacy>=3.5.0

# Computer Vision
opencv-python>=4.8.0

# Utils
tqdm>=4.65.0